// src/types/generated/inventoryControllerFindAllParams.ts
/**
 * Generated by orval v8.1.0 🍺
 * Do not edit manually.
 * Inventory Management API
 * A comprehensive NestJS inventory and order management system with role-based access control, stock tracking, and audit logging.
 * OpenAPI spec version: 1.0
 */

export type InventoryControllerFindAllParams = {
page?: number;
limit?: number;
};