// src/types/generated/paginatedInventoryDataDto.ts
/**
 * Generated by orval v8.1.0 🍺
 * Do not edit manually.
 * Inventory Management API
 * A comprehensive NestJS inventory and order management system with role-based access control, stock tracking, and audit logging.
 * OpenAPI spec version: 1.0
 */
import type { InventoryResponseDto } from './inventoryResponseDto';

export interface PaginatedInventoryDataDto {
  items: InventoryResponseDto[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}